feat(gpu): signed addition
