+++ tfhe/src/integer/gpu/server_key/radix/mod.rs
@@ -5,7 +5,9 @@ use crate::core_crypto::gpu::CudaStream;
 use crate::core_crypto::prelude::{ContiguousEntityContainerMut, LweCiphertextCount};
 use crate::integer::block_decomposition::{BlockDecomposer, DecomposableInto};
 use crate::integer::gpu::ciphertext::info::{CudaBlockInfo, CudaRadixCiphertextInfo};
-use crate::integer::gpu::ciphertext::CudaRadixCiphertext;
+use crate::integer::gpu::ciphertext::{
+    CudaIntegerRadixCiphertext, CudaRadixCiphertext, CudaUnsignedRadixCiphertext,
+};
 use crate::integer::gpu::server_key::CudaBootstrappingKey;
 use crate::integer::gpu::CudaServerKey;
 use crate::shortint::ciphertext::{Degree, NoiseLevel};
@@ -28,14 +30,18 @@ mod sub;
 mod scalar_rotate;
 #[cfg(test)]
 mod tests;
+#[cfg(test)]
+mod tests_signed;
+#[cfg(test)]
+mod tests_unsigned;
+
 impl CudaServerKey {
-    /// Create a trivial ciphertext filled with zeros
+    /// Create a trivial ciphertext filled with zeros on the GPU.
     ///
     /// # Example
     ///
     /// ```rust
     /// use tfhe::core_crypto::gpu::{CudaDevice, CudaStream};
-    /// use tfhe::integer::gpu::ciphertext::CudaRadixCiphertext;
     /// use tfhe::integer::gpu::gen_keys_radix_gpu;
     /// use tfhe::integer::{gen_keys_radix, RadixCiphertext};
     /// use tfhe::shortint::parameters::PARAM_MESSAGE_2_CARRY_2_KS_PBS;
@@ -49,7 +55,7 @@ impl CudaServerKey {
     /// // Generate the client key and the server key:
     /// let (cks, sks) = gen_keys_radix_gpu(PARAM_MESSAGE_2_CARRY_2_KS_PBS, num_blocks, &mut stream);
     ///
-    /// let d_ctxt: CudaRadixCiphertext = sks.create_trivial_zero_radix(num_blocks, &mut stream);
+    /// let d_ctxt = sks.create_trivial_zero_radix(num_blocks, &mut stream);
     /// let ctxt = d_ctxt.to_radix_ciphertext(&mut stream);
     ///
     /// // Decrypt:
@@ -60,17 +66,16 @@ impl CudaServerKey {
         &self,
         num_blocks: usize,
         stream: &CudaStream,
-    ) -> CudaRadixCiphertext {
+    ) -> CudaUnsignedRadixCiphertext {
         self.create_trivial_radix(0, num_blocks, stream)
     }
 
-    /// Create a trivial ciphertext
+    /// Create a trivial ciphertext on the GPU
     ///
     /// # Example
     ///
     /// ```rust
     /// use tfhe::core_crypto::gpu::{CudaDevice, CudaStream};
-    /// use tfhe::integer::gpu::ciphertext::CudaRadixCiphertext;
     /// use tfhe::integer::gpu::gen_keys_radix_gpu;
     /// use tfhe::integer::{gen_keys_radix, RadixCiphertext};
     /// use tfhe::shortint::parameters::PARAM_MESSAGE_2_CARRY_2_KS_PBS;
@@ -84,21 +89,21 @@ impl CudaServerKey {
     /// // Generate the client key and the server key:
     /// let (cks, sks) = gen_keys_radix_gpu(PARAM_MESSAGE_2_CARRY_2_KS_PBS, num_blocks, &mut stream);
     ///
-    /// let d_ctxt: CudaRadixCiphertext = sks.create_trivial_radix(212u64, num_blocks, &mut stream);
+    /// let d_ctxt = sks.create_trivial_radix(212u64, num_blocks, &mut stream);
     /// let ctxt = d_ctxt.to_radix_ciphertext(&mut stream);
     ///
     /// // Decrypt:
     /// let dec: u64 = cks.decrypt(&ctxt);
     /// assert_eq!(212, dec);
     /// ```
-    pub fn create_trivial_radix<T>(
+    pub fn create_trivial_radix<Scalar>(
         &self,
-        scalar: T,
+        scalar: Scalar,
         num_blocks: usize,
         stream: &CudaStream,
-    ) -> CudaRadixCiphertext
+    ) -> CudaUnsignedRadixCiphertext
     where
-        T: DecomposableInto<u64>,
+        Scalar: DecomposableInto<u64>,
     {
         let lwe_size = match self.pbs_order {
             PBSOrder::KeyswitchBootstrap => self.key_switching_key.input_key_lwe_size(),
@@ -131,9 +136,11 @@ impl CudaServerKey {
 
         let d_blocks = CudaLweCiphertextList::from_lwe_ciphertext_list(&cpu_lwe_list, stream);
 
-        CudaRadixCiphertext {
-            d_blocks,
-            info: CudaRadixCiphertextInfo { blocks: info },
+        CudaUnsignedRadixCiphertext {
+            ciphertext: CudaRadixCiphertext {
+                d_blocks,
+                info: CudaRadixCiphertextInfo { blocks: info },
+            },
         }
     }
 
@@ -141,16 +148,19 @@ impl CudaServerKey {
     ///
     /// - `stream` __must__ be synchronized to guarantee computation has finished, and inputs must
     ///   not be dropped until stream is synchronized
-    pub(crate) unsafe fn propagate_single_carry_assign_async(
+    pub(crate) unsafe fn propagate_single_carry_assign_async<T>(
         &self,
-        ct: &mut CudaRadixCiphertext,
+        ct: &mut T,
         stream: &CudaStream,
-    ) {
-        let num_blocks = ct.d_blocks.lwe_ciphertext_count().0 as u32;
+    ) where
+        T: CudaIntegerRadixCiphertext,
+    {
+        let ciphertext = ct.as_mut();
+        let num_blocks = ciphertext.d_blocks.lwe_ciphertext_count().0 as u32;
         match &self.bootstrapping_key {
             CudaBootstrappingKey::Classic(d_bsk) => {
                 stream.propagate_single_carry_classic_assign_async(
-                    &mut ct.d_blocks.0.d_vec,
+                    &mut ciphertext.d_blocks.0.d_vec,
                     &d_bsk.d_vec,
                     &self.key_switching_key.d_vec,
                     d_bsk.input_lwe_dimension(),
@@ -161,13 +171,13 @@ impl CudaServerKey {
                     d_bsk.decomp_level_count(),
                     d_bsk.decomp_base_log(),
                     num_blocks,
-                    ct.info.blocks.first().unwrap().message_modulus,
-                    ct.info.blocks.first().unwrap().carry_modulus,
+                    ciphertext.info.blocks.first().unwrap().message_modulus,
+                    ciphertext.info.blocks.first().unwrap().carry_modulus,
                 );
             }
             CudaBootstrappingKey::MultiBit(d_multibit_bsk) => {
                 stream.propagate_single_carry_multibit_assign_async(
-                    &mut ct.d_blocks.0.d_vec,
+                    &mut ciphertext.d_blocks.0.d_vec,
                     &d_multibit_bsk.d_vec,
                     &self.key_switching_key.d_vec,
                     d_multibit_bsk.input_lwe_dimension(),
@@ -179,12 +189,12 @@ impl CudaServerKey {
                     d_multibit_bsk.decomp_base_log(),
                     d_multibit_bsk.grouping_factor,
                     num_blocks,
-                    ct.info.blocks.first().unwrap().message_modulus,
-                    ct.info.blocks.first().unwrap().carry_modulus,
+                    ciphertext.info.blocks.first().unwrap().message_modulus,
+                    ciphertext.info.blocks.first().unwrap().carry_modulus,
                 );
             }
         };
-        ct.info.blocks.iter_mut().for_each(|b| {
+        ciphertext.info.blocks.iter_mut().for_each(|b| {
             b.degree = Degree::new(b.message_modulus.0 - 1);
             b.noise_level = NoiseLevel::NOMINAL;
         });
@@ -194,16 +204,17 @@ impl CudaServerKey {
     ///
     /// - `stream` __must__ be synchronized to guarantee computation has finished, and inputs must
     ///   not be dropped until stream is synchronized
-    pub(crate) unsafe fn full_propagate_assign_async(
+    pub(crate) unsafe fn full_propagate_assign_async<T: CudaIntegerRadixCiphertext>(
         &self,
-        ct: &mut CudaRadixCiphertext,
+        ct: &mut T,
         stream: &CudaStream,
     ) {
-        let num_blocks = ct.d_blocks.lwe_ciphertext_count().0 as u32;
+        let ciphertext = ct.as_mut();
+        let num_blocks = ciphertext.d_blocks.lwe_ciphertext_count().0 as u32;
         match &self.bootstrapping_key {
             CudaBootstrappingKey::Classic(d_bsk) => {
                 stream.full_propagate_classic_assign_async(
-                    &mut ct.d_blocks.0.d_vec,
+                    &mut ciphertext.d_blocks.0.d_vec,
                     &d_bsk.d_vec,
                     &self.key_switching_key.d_vec,
                     d_bsk.input_lwe_dimension(),
@@ -214,13 +225,13 @@ impl CudaServerKey {
                     d_bsk.decomp_level_count(),
                     d_bsk.decomp_base_log(),
                     num_blocks,
-                    ct.info.blocks.first().unwrap().message_modulus,
-                    ct.info.blocks.first().unwrap().carry_modulus,
+                    ciphertext.info.blocks.first().unwrap().message_modulus,
+                    ciphertext.info.blocks.first().unwrap().carry_modulus,
                 );
             }
             CudaBootstrappingKey::MultiBit(d_multibit_bsk) => {
                 stream.full_propagate_multibit_assign_async(
-                    &mut ct.d_blocks.0.d_vec,
+                    &mut ciphertext.d_blocks.0.d_vec,
                     &d_multibit_bsk.d_vec,
                     &self.key_switching_key.d_vec,
                     d_multibit_bsk.input_lwe_dimension(),
@@ -232,19 +243,21 @@ impl CudaServerKey {
                     d_multibit_bsk.decomp_base_log(),
                     d_multibit_bsk.grouping_factor,
                     num_blocks,
-                    ct.info.blocks.first().unwrap().message_modulus,
-                    ct.info.blocks.first().unwrap().carry_modulus,
+                    ciphertext.info.blocks.first().unwrap().message_modulus,
+                    ciphertext.info.blocks.first().unwrap().carry_modulus,
                 );
             }
         };
-        ct.info
+        ciphertext
+            .info
             .blocks
             .iter_mut()
             .for_each(|b| b.degree = Degree::new(b.message_modulus.0 - 1));
     }
 
-    /// Prepend trivial zero LSB blocks to an existing [`CudaRadixCiphertext`] and returns the
-    /// result as a new [`CudaRadixCiphertext`]. This can be useful for casting operations.
+    /// Prepend trivial zero LSB blocks to an existing [`CudaUnsignedRadixCiphertext`] or
+    /// [`CudaSignedRadixCiphertext`] and returns the result as a new ciphertext on GPU.
+    /// This can be useful for casting operations.
     ///
     /// # Example
     ///
@@ -264,13 +277,12 @@ impl CudaServerKey {
     /// // Generate the client key and the server key:
     /// let (cks, sks) = gen_keys_radix_gpu(PARAM_MESSAGE_2_CARRY_2_KS_PBS, num_blocks, &mut stream);
     ///
-    /// let mut d_ct1: CudaRadixCiphertext = sks.create_trivial_radix(7u64, num_blocks, &mut stream);
+    /// let mut d_ct1 = sks.create_trivial_radix(7u64, num_blocks, &mut stream);
     /// let ct1 = d_ct1.to_radix_ciphertext(&mut stream);
     /// assert_eq!(ct1.blocks().len(), 4);
     ///
     /// let added_blocks = 2;
-    /// let d_ct_res =
-    ///     sks.extend_radix_with_trivial_zero_blocks_lsb(&mut d_ct1, added_blocks, &mut stream);
+    /// let d_ct_res = sks.extend_radix_with_trivial_zero_blocks_lsb(&d_ct1, added_blocks, &mut stream);
     /// let ct_res = d_ct_res.to_radix_ciphertext(&mut stream);
     /// assert_eq!(ct_res.blocks().len(), 6);
     ///
@@ -281,22 +293,26 @@ impl CudaServerKey {
     ///     res
     /// );
     /// ```
-    pub fn extend_radix_with_trivial_zero_blocks_lsb(
+    pub fn extend_radix_with_trivial_zero_blocks_lsb<T: CudaIntegerRadixCiphertext>(
         &self,
-        ct: &CudaRadixCiphertext,
+        ct: &T,
         num_blocks: usize,
         stream: &CudaStream,
-    ) -> CudaRadixCiphertext {
-        let new_num_blocks = ct.d_blocks.lwe_ciphertext_count().0 + num_blocks;
-        let ciphertext_modulus = ct.d_blocks.ciphertext_modulus();
-        let lwe_size = ct.d_blocks.lwe_dimension().to_lwe_size();
+    ) -> T {
+        let new_num_blocks = ct.as_ref().d_blocks.lwe_ciphertext_count().0 + num_blocks;
+        let ciphertext_modulus = ct.as_ref().d_blocks.ciphertext_modulus();
+        let lwe_size = ct.as_ref().d_blocks.lwe_dimension().to_lwe_size();
         let shift = num_blocks * lwe_size.0;
 
         let mut extended_ct_vec =
             unsafe { CudaVec::new_async(new_num_blocks * lwe_size.0, stream) };
         unsafe {
             extended_ct_vec.memset_async(0u64, stream);
-            extended_ct_vec.copy_self_range_gpu_to_gpu_async(shift.., &ct.d_blocks.0.d_vec, stream);
+            extended_ct_vec.copy_self_range_gpu_to_gpu_async(
+                shift..,
+                &ct.as_ref().d_blocks.0.d_vec,
+                stream,
+            );
         }
         stream.synchronize();
         let extended_ct_list = CudaLweCiphertextList::from_cuda_vec(
@@ -306,19 +322,20 @@ impl CudaServerKey {
         );
 
         let extended_ct_info = ct
+            .as_ref()
             .info
             .after_extend_radix_with_trivial_zero_blocks_lsb(num_blocks);
-        CudaRadixCiphertext::new(extended_ct_list, extended_ct_info)
+        T::from(CudaRadixCiphertext::new(extended_ct_list, extended_ct_info))
     }
 
-    /// Append trivial zero MSB blocks to an existing [`CudaRadixCiphertext`] and returns the result
-    /// as a new [`CudaRadixCiphertext`]. This can be useful for casting operations.
+    /// Append trivial zero MSB blocks to an existing [`CudaUnsignedRadixCiphertext`] or
+    /// [`CudaSignedRadixCiphertext`] and returns the result as a new ciphertext on GPU.
+    /// This can be useful for casting operations.
     ///
     /// # Example
     ///
     ///```rust
     /// use tfhe::core_crypto::gpu::{CudaDevice, CudaStream};
-    /// use tfhe::integer::gpu::ciphertext::CudaRadixCiphertext;
     /// use tfhe::integer::gpu::gen_keys_radix_gpu;
     /// use tfhe::integer::IntegerCiphertext;
     /// use tfhe::shortint::parameters::PARAM_MESSAGE_2_CARRY_2_KS_PBS;
@@ -332,7 +349,7 @@ impl CudaServerKey {
     /// // Generate the client key and the server key:
     /// let (cks, sks) = gen_keys_radix_gpu(PARAM_MESSAGE_2_CARRY_2_KS_PBS, num_blocks, &mut stream);
     ///
-    /// let mut d_ct1: CudaRadixCiphertext = sks.create_trivial_radix(7u64, num_blocks, &mut stream);
+    /// let mut d_ct1 = sks.create_trivial_radix(7u64, num_blocks, &mut stream);
     /// let ct1 = d_ct1.to_radix_ciphertext(&mut stream);
     /// assert_eq!(ct1.blocks().len(), 4);
     ///
@@ -344,21 +361,21 @@ impl CudaServerKey {
     /// let res: u64 = cks.decrypt(&ct_res);
     /// assert_eq!(7, res);
     /// ```
-    pub fn extend_radix_with_trivial_zero_blocks_msb(
+    pub fn extend_radix_with_trivial_zero_blocks_msb<T: CudaIntegerRadixCiphertext>(
         &self,
-        ct: &CudaRadixCiphertext,
+        ct: &T,
         num_blocks: usize,
         stream: &CudaStream,
-    ) -> CudaRadixCiphertext {
-        let new_num_blocks = ct.d_blocks.lwe_ciphertext_count().0 + num_blocks;
-        let ciphertext_modulus = ct.d_blocks.ciphertext_modulus();
-        let lwe_size = ct.d_blocks.lwe_dimension().to_lwe_size();
+    ) -> T {
+        let new_num_blocks = ct.as_ref().d_blocks.lwe_ciphertext_count().0 + num_blocks;
+        let ciphertext_modulus = ct.as_ref().d_blocks.ciphertext_modulus();
+        let lwe_size = ct.as_ref().d_blocks.lwe_dimension().to_lwe_size();
 
         let mut extended_ct_vec =
             unsafe { CudaVec::new_async(new_num_blocks * lwe_size.0, stream) };
         unsafe {
             extended_ct_vec.memset_async(0u64, stream);
-            extended_ct_vec.copy_from_gpu_async(&ct.d_blocks.0.d_vec, stream);
+            extended_ct_vec.copy_from_gpu_async(&ct.as_ref().d_blocks.0.d_vec, stream);
         }
         stream.synchronize();
         let extended_ct_list = CudaLweCiphertextList::from_cuda_vec(
@@ -368,19 +385,20 @@ impl CudaServerKey {
         );
 
         let extended_ct_info = ct
+            .as_ref()
             .info
             .after_extend_radix_with_trivial_zero_blocks_msb(num_blocks);
-        CudaRadixCiphertext::new(extended_ct_list, extended_ct_info)
+        T::from(CudaRadixCiphertext::new(extended_ct_list, extended_ct_info))
     }
 
-    /// Remove LSB blocks from an existing [`CudaRadixCiphertext`] and returns the result as a new
-    /// [`CudaRadixCiphertext`]. This can be useful for casting operations.
+    /// Remove LSB blocks from an existing [`CudaUnsignedRadixCiphertext`] or
+    /// [`CudaSignedRadixCiphertext`] and returns the result as a new ciphertext on GPU.
+    /// This can be useful for casting operations.
     ///
     /// # Example
     ///
     ///```rust
     /// use tfhe::core_crypto::gpu::{CudaDevice, CudaStream};
-    /// use tfhe::integer::gpu::ciphertext::CudaRadixCiphertext;
     /// use tfhe::integer::gpu::gen_keys_radix_gpu;
     /// use tfhe::integer::IntegerCiphertext;
     /// use tfhe::shortint::parameters::PARAM_MESSAGE_2_CARRY_2_KS_PBS;
@@ -394,7 +412,7 @@ impl CudaServerKey {
     /// // Generate the client key and the server key:
     /// let (cks, sks) = gen_keys_radix_gpu(PARAM_MESSAGE_2_CARRY_2_KS_PBS, num_blocks, &mut stream);
     ///
-    /// let mut d_ct1: CudaRadixCiphertext = sks.create_trivial_radix(119u64, num_blocks, &mut stream);
+    /// let mut d_ct1 = sks.create_trivial_radix(119u64, num_blocks, &mut stream);
     /// let ct1 = d_ct1.to_radix_ciphertext(&mut stream);
     /// assert_eq!(ct1.blocks().len(), 4);
     ///
@@ -406,20 +424,24 @@ impl CudaServerKey {
     /// let res: u64 = cks.decrypt(&ct_res);
     /// assert_eq!(7, res);
     /// ```
-    pub fn trim_radix_blocks_lsb(
+    pub fn trim_radix_blocks_lsb<T: CudaIntegerRadixCiphertext>(
         &self,
-        ct: &CudaRadixCiphertext,
+        ct: &T,
         num_blocks: usize,
         stream: &CudaStream,
-    ) -> CudaRadixCiphertext {
-        let new_num_blocks = ct.d_blocks.lwe_ciphertext_count().0 - num_blocks;
-        let ciphertext_modulus = ct.d_blocks.ciphertext_modulus();
-        let lwe_size = ct.d_blocks.lwe_dimension().to_lwe_size();
+    ) -> T {
+        let new_num_blocks = ct.as_ref().d_blocks.lwe_ciphertext_count().0 - num_blocks;
+        let ciphertext_modulus = ct.as_ref().d_blocks.ciphertext_modulus();
+        let lwe_size = ct.as_ref().d_blocks.lwe_dimension().to_lwe_size();
         let shift = num_blocks * lwe_size.0;
 
         let mut trimmed_ct_vec = unsafe { CudaVec::new_async(new_num_blocks * lwe_size.0, stream) };
         unsafe {
-            trimmed_ct_vec.copy_src_range_gpu_to_gpu_async(shift.., &ct.d_blocks.0.d_vec, stream);
+            trimmed_ct_vec.copy_src_range_gpu_to_gpu_async(
+                shift..,
+                &ct.as_ref().d_blocks.0.d_vec,
+                stream,
+            );
         }
         stream.synchronize();
         let trimmed_ct_list = CudaLweCiphertextList::from_cuda_vec(
@@ -428,18 +450,18 @@ impl CudaServerKey {
             ciphertext_modulus,
         );
 
-        let trimmed_ct_info = ct.info.after_trim_radix_blocks_lsb(num_blocks);
-        CudaRadixCiphertext::new(trimmed_ct_list, trimmed_ct_info)
+        let trimmed_ct_info = ct.as_ref().info.after_trim_radix_blocks_lsb(num_blocks);
+        T::from(CudaRadixCiphertext::new(trimmed_ct_list, trimmed_ct_info))
     }
 
-    /// Remove MSB blocks from an existing [`CudaRadixCiphertext`] and returns the result as a new
-    /// [`CudaRadixCiphertext`]. This can be useful for casting operations.
+    /// Remove MSB blocks from an existing [`CudaUnsignedRadixCiphertext`] or
+    /// [`CudaSignedRadixCiphertext`] and returns the result as a new ciphertext on GPU.
+    /// This can be useful for casting operations.
     ///
     /// # Example
     ///
     ///```rust
     /// use tfhe::core_crypto::gpu::{CudaDevice, CudaStream};
-    /// use tfhe::integer::gpu::ciphertext::CudaRadixCiphertext;
     /// use tfhe::integer::gpu::gen_keys_radix_gpu;
     /// use tfhe::integer::IntegerCiphertext;
     /// use tfhe::shortint::parameters::PARAM_MESSAGE_2_CARRY_2_KS_PBS;
@@ -453,7 +475,7 @@ impl CudaServerKey {
     /// // Generate the client key and the server key:
     /// let (cks, sks) = gen_keys_radix_gpu(PARAM_MESSAGE_2_CARRY_2_KS_PBS, num_blocks, &mut stream);
     ///
-    /// let mut d_ct1: CudaRadixCiphertext = sks.create_trivial_radix(119u64, num_blocks, &mut stream);
+    /// let mut d_ct1 = sks.create_trivial_radix(119u64, num_blocks, &mut stream);
     /// let ct1 = d_ct1.to_radix_ciphertext(&mut stream);
     /// assert_eq!(ct1.blocks().len(), 4);
     ///
@@ -465,20 +487,24 @@ impl CudaServerKey {
     /// let res: u64 = cks.decrypt(&ct_res);
     /// assert_eq!(7, res);
     /// ```
-    pub fn trim_radix_blocks_msb(
+    pub fn trim_radix_blocks_msb<T: CudaIntegerRadixCiphertext>(
         &self,
-        ct: &CudaRadixCiphertext,
+        ct: &T,
         num_blocks: usize,
         stream: &CudaStream,
-    ) -> CudaRadixCiphertext {
-        let new_num_blocks = ct.d_blocks.lwe_ciphertext_count().0 - num_blocks;
-        let ciphertext_modulus = ct.d_blocks.ciphertext_modulus();
-        let lwe_size = ct.d_blocks.lwe_dimension().to_lwe_size();
+    ) -> T {
+        let new_num_blocks = ct.as_ref().d_blocks.lwe_ciphertext_count().0 - num_blocks;
+        let ciphertext_modulus = ct.as_ref().d_blocks.ciphertext_modulus();
+        let lwe_size = ct.as_ref().d_blocks.lwe_dimension().to_lwe_size();
         let shift = new_num_blocks * lwe_size.0;
 
         let mut trimmed_ct_vec = unsafe { CudaVec::new_async(new_num_blocks * lwe_size.0, stream) };
         unsafe {
-            trimmed_ct_vec.copy_src_range_gpu_to_gpu_async(0..shift, &ct.d_blocks.0.d_vec, stream);
+            trimmed_ct_vec.copy_src_range_gpu_to_gpu_async(
+                0..shift,
+                &ct.as_ref().d_blocks.0.d_vec,
+                stream,
+            );
         }
         stream.synchronize();
         let trimmed_ct_list = CudaLweCiphertextList::from_cuda_vec(
@@ -487,18 +513,17 @@ impl CudaServerKey {
             ciphertext_modulus,
         );
 
-        let trimmed_ct_info = ct.info.after_trim_radix_blocks_msb(num_blocks);
-        CudaRadixCiphertext::new(trimmed_ct_list, trimmed_ct_info)
+        let trimmed_ct_info = ct.as_ref().info.after_trim_radix_blocks_msb(num_blocks);
+        T::from(CudaRadixCiphertext::new(trimmed_ct_list, trimmed_ct_info))
     }
 
-    /// Cast a CudaRadixCiphertext to a CudaRadixCiphertext
+    /// Cast a [`CudaUnsignedRadixCiphertext`] to a [`CudaUnsignedRadixCiphertext`]
     /// with a possibly different number of blocks
     ///
     /// # Example
     ///
     ///```rust
     /// use tfhe::core_crypto::gpu::{CudaDevice, CudaStream};
-    /// use tfhe::integer::gpu::ciphertext::CudaRadixCiphertext;
     /// use tfhe::integer::gpu::gen_keys_radix_gpu;
     /// use tfhe::integer::IntegerCiphertext;
     /// use tfhe::shortint::parameters::PARAM_MESSAGE_2_CARRY_2_KS_PBS;
@@ -513,7 +538,7 @@ impl CudaServerKey {
     ///
     /// let msg = 2u8;
     ///
-    /// let mut d_ct1: CudaRadixCiphertext = sks.create_trivial_radix(msg, num_blocks, &mut stream);
+    /// let mut d_ct1 = sks.create_trivial_radix(msg, num_blocks, &mut stream);
     /// let ct1 = d_ct1.to_radix_ciphertext(&mut stream);
     /// assert_eq!(ct1.blocks().len(), 4);
     ///
@@ -527,17 +552,17 @@ impl CudaServerKey {
     /// ```
     pub fn cast_to_unsigned(
         &self,
-        mut source: CudaRadixCiphertext,
+        mut source: CudaUnsignedRadixCiphertext,
         target_num_blocks: usize,
         stream: &CudaStream,
-    ) -> CudaRadixCiphertext {
+    ) -> CudaUnsignedRadixCiphertext {
         if !source.block_carries_are_empty() {
             unsafe {
                 self.full_propagate_assign_async(&mut source, stream);
             }
             stream.synchronize();
         }
-        let current_num_blocks = source.info.blocks.len();
+        let current_num_blocks = source.ciphertext.info.blocks.len();
         // Casting from unsigned to unsigned, this is just about trimming/extending with zeros
         if target_num_blocks > current_num_blocks {
             let num_blocks_to_add = target_num_blocks - current_num_blocks;
