+++ tfhe/src/integer/gpu/server_key/radix/scalar_mul.rs
@@ -1,5 +1,5 @@
 use crate::core_crypto::gpu::CudaStream;
-use crate::integer::gpu::ciphertext::CudaRadixCiphertext;
+use crate::integer::gpu::ciphertext::{CudaIntegerRadixCiphertext, CudaUnsignedRadixCiphertext};
 use crate::integer::gpu::server_key::CudaServerKey;
 
 impl CudaServerKey {
@@ -15,7 +15,7 @@ impl CudaServerKey {
     /// ```rust
     /// use tfhe::core_crypto::gpu::{CudaDevice, CudaStream};
     /// use tfhe::integer::gen_keys_radix;
-    /// use tfhe::integer::gpu::ciphertext::CudaRadixCiphertext;
+    /// use tfhe::integer::gpu::ciphertext::CudaUnsignedRadixCiphertext;
     /// use tfhe::integer::gpu::gen_keys_radix_gpu;
     /// use tfhe::shortint::parameters::PARAM_MESSAGE_2_CARRY_2_KS_PBS;
     ///
@@ -31,7 +31,7 @@ impl CudaServerKey {
     /// let scalar = 3;
     ///
     /// let ct = cks.encrypt(msg);
-    /// let mut d_ct = CudaRadixCiphertext::from_radix_ciphertext(&ct, &mut stream);
+    /// let mut d_ct = CudaUnsignedRadixCiphertext::from_radix_ciphertext(&ct, &mut stream);
     ///
     /// // Compute homomorphically a scalar multiplication:
     /// let d_ct_res = sks.unchecked_small_scalar_mul(&d_ct, scalar, &mut stream);
@@ -42,10 +42,10 @@ impl CudaServerKey {
     /// ```
     pub fn unchecked_small_scalar_mul(
         &self,
-        ct: &CudaRadixCiphertext,
+        ct: &CudaUnsignedRadixCiphertext,
         scalar: u64,
         stream: &CudaStream,
-    ) -> CudaRadixCiphertext {
+    ) -> CudaUnsignedRadixCiphertext {
         let mut result = unsafe { ct.duplicate_async(stream) };
         self.unchecked_small_scalar_mul_assign(&mut result, scalar, stream);
         result
@@ -57,23 +57,23 @@ impl CudaServerKey {
     ///   not be dropped until stream is synchronised
     pub unsafe fn unchecked_small_scalar_mul_assign_async(
         &self,
-        ct: &mut CudaRadixCiphertext,
+        ct: &mut CudaUnsignedRadixCiphertext,
         scalar: u64,
         stream: &CudaStream,
     ) {
         match scalar {
             0 => {
-                ct.d_blocks.0.d_vec.memset_async(0, stream);
+                ct.as_mut().d_blocks.0.d_vec.memset_async(0, stream);
             }
             1 => {
                 // Multiplication by one is the identity
             }
             _ => {
-                let lwe_dimension = ct.d_blocks.lwe_dimension();
-                let lwe_ciphertext_count = ct.d_blocks.lwe_ciphertext_count();
+                let lwe_dimension = ct.as_ref().d_blocks.lwe_dimension();
+                let lwe_ciphertext_count = ct.as_ref().d_blocks.lwe_ciphertext_count();
 
                 stream.small_scalar_mult_integer_radix_assign_async(
-                    &mut ct.d_blocks.0.d_vec,
+                    &mut ct.as_mut().d_blocks.0.d_vec,
                     scalar,
                     lwe_dimension,
                     lwe_ciphertext_count.0 as u32,
@@ -81,12 +81,12 @@ impl CudaServerKey {
             }
         }
 
-        ct.info = ct.info.after_small_scalar_mul(scalar as u8);
+        ct.as_mut().info = ct.as_ref().info.after_small_scalar_mul(scalar as u8);
     }
 
     pub fn unchecked_small_scalar_mul_assign(
         &self,
-        ct: &mut CudaRadixCiphertext,
+        ct: &mut CudaUnsignedRadixCiphertext,
         scalar: u64,
         stream: &CudaStream,
     ) {
@@ -108,7 +108,7 @@ impl CudaServerKey {
     /// ```rust
     /// use tfhe::core_crypto::gpu::{CudaDevice, CudaStream};
     /// use tfhe::integer::gen_keys_radix;
-    /// use tfhe::integer::gpu::ciphertext::CudaRadixCiphertext;
+    /// use tfhe::integer::gpu::ciphertext::CudaUnsignedRadixCiphertext;
     /// use tfhe::integer::gpu::gen_keys_radix_gpu;
     /// use tfhe::shortint::parameters::PARAM_MESSAGE_2_CARRY_2_KS_PBS;
     ///
@@ -124,7 +124,7 @@ impl CudaServerKey {
     /// let scalar = 3;
     ///
     /// let ct = cks.encrypt(msg);
-    /// let mut d_ct = CudaRadixCiphertext::from_radix_ciphertext(&ct, &mut stream);
+    /// let mut d_ct = CudaUnsignedRadixCiphertext::from_radix_ciphertext(&ct, &mut stream);
     ///
     /// // Compute homomorphically a scalar multiplication:
     /// let d_ct_res = sks.small_scalar_mul(&d_ct, scalar, &mut stream);
@@ -135,10 +135,10 @@ impl CudaServerKey {
     /// ```
     pub fn small_scalar_mul(
         &self,
-        ct: &CudaRadixCiphertext,
+        ct: &CudaUnsignedRadixCiphertext,
         scalar: u64,
         stream: &CudaStream,
-    ) -> CudaRadixCiphertext {
+    ) -> CudaUnsignedRadixCiphertext {
         let mut result = unsafe { ct.duplicate_async(stream) };
         self.small_scalar_mul_assign(&mut result, scalar, stream);
         result
@@ -150,7 +150,7 @@ impl CudaServerKey {
     ///   not be dropped until stream is synchronised
     pub unsafe fn small_scalar_mul_assign_async(
         &self,
-        ct: &mut CudaRadixCiphertext,
+        ct: &mut CudaUnsignedRadixCiphertext,
         scalar: u64,
         stream: &CudaStream,
     ) {
@@ -164,7 +164,7 @@ impl CudaServerKey {
 
     pub fn small_scalar_mul_assign(
         &self,
-        ct: &mut CudaRadixCiphertext,
+        ct: &mut CudaUnsignedRadixCiphertext,
         scalar: u64,
         stream: &CudaStream,
     ) {
