+++ tfhe/src/integer/server_key/radix_parallel/ilog2.rs
@@ -1110,9 +1110,10 @@ pub(crate) mod tests_unsigned {
 pub(crate) mod tests_signed {
     use super::*;
     use crate::integer::keycache::KEY_CACHE;
-    use crate::integer::server_key::radix_parallel::tests_signed::{
-        random_non_zero_value, signed_add_under_modulus, NB_CTXT, NB_TESTS_SMALLER,
+    use crate::integer::server_key::radix_parallel::tests_cases_signed::{
+        random_non_zero_value, signed_add_under_modulus,
     };
+    use crate::integer::server_key::radix_parallel::tests_signed::{NB_CTXT, NB_TESTS_SMALLER};
     use crate::integer::{IntegerKeyKind, RadixClientKey, SignedRadixCiphertext};
     use crate::shortint::PBSParameters;
     use rand::Rng;
